week5/project/asbolsyn/requirements.txt
astroid==2.0.4
Babel==2.7.0
Django==2.1.1
django-bootstrap3==11.0.0
django-cors-headers==3.1.0
django-crispy-forms==1.7.2
django-datetimepicker==3.14
django-extensions==2.1.2
django-filter==2.0.0
django-guardian==1.4.9
django-phonenumber-field==3.0.1
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
phonenumbers==8.10.17
psycopg2==2.8.3
pydotplus==2.0.2
PyJWT==1.7.1
pylint==2.1.1
pyparsing==2.2.1
pytz==2018.5
six==1.11.0
typed-ast==1.1.0
wrapt==1.10.11
